Stream gsplat flatten into a single leaf (#2348) (#2359)
* fix(gsplats): stream flatten into a single leaf (#2348)

* fix(encoding): keep precision staging zero-copy (#2348)

* fix(gsplats): preserve streaming flatten invariants (#2348)

* fix(gsplats): preserve flatten metadata invariants

* fix(gsplats): bound streaming flatten memory

* docs(gsplats): document streaming flatten tradeoffs

* fix(gsplats): preserve barriers in streaming flatten

* fix(gsplats): restore flatten writer contracts

* fix(gsplats): harden streaming flatten

* fix(gsplats): cap global barrier detection

* fix(gsplats): align streaming barrier detection

* refactor(gsplats): keep barrier helpers internal

* fix(gsplats): bound streaming barrier runs

* docs(gsplats): restore streaming writer contracts

* test(gsplats): cover exact barrier run overflow

* refactor(gsplats): keep barrier guard below complexity limit

* fix(gsplats): allow large explicit barrier stacks

---------

+++ packages/luxar/src/luxar/encoding/_encoders/base.py
@@ -474,7 +474,7 @@ class BaseEncoderMixin:
         create_array(
             zarr_group,
             name,
-            data=data.astype(target_dtype),
+            data=data.astype(target_dtype, copy=False),
             chunks=chunks,
             compressor=resolve_compressor(compressor, target_dtype),
             overwrite=True,
